+++ src/xulbux/base/consts.py
@@ -97,8 +97,12 @@ class ANSI:
     SEQ_BG_COLOR: Final[FormattableString] = CHAR + START + "48" + SEP + "2" + SEP + "{}" + SEP + "{}" + SEP + "{}" + END
     """ANSI escape sequence with three placeholders for setting the RGB background color."""
 
-    COLOR_MAP: Final[tuple[str, ...]] = (
-        ########### DEFAULT CONSOLE COLOR NAMES ############
+    SEQ_LINK_OPEN: Final[FormattableString] = CHAR + "]8;;{}" + CHAR + "\\"
+    """OSC 8 hyperlink opening sequence with a placeholder for the URL."""
+    SEQ_LINK_CLOSE: Final[str] = CHAR + "]8;;" + CHAR + "\\"
+    """OSC 8 hyperlink closing sequence."""
+
+    COLOR_MAP: Final[set[str]] = {
         "black",
         "red",
         "green",
@@ -107,8 +111,28 @@ class ANSI:
         "magenta",
         "cyan",
         "white",
-    )
+    }
     """The standard terminal color names."""
+
+    COLOR_VARIANTS_MAP: Final[set[str]] = COLOR_MAP | {
+        "br:black",
+        "br:red",
+        "br:green",
+        "br:yellow",
+        "br:blue",
+        "br:magenta",
+        "br:cyan",
+        "br:white",
+        "bright:black",
+        "bright:red",
+        "bright:green",
+        "bright:yellow",
+        "bright:blue",
+        "bright:magenta",
+        "bright:cyan",
+        "bright:white",
+    }
+    """All color variants that can be used in formatting."""
 
     CODES_MAP: Final[dict[str | tuple[str, ...], int]] = {
         ################# SPECIFIC RESETS ##################